chore: v2 적용 전 기존 코드 정리 (#47)
