+++ src/App.tsx
@@ -3,7 +3,7 @@ import { Route, Routes } from 'react-router-dom';
 import LoginPage from './page/Login/LoginPage';
 import HomeMyworkPage from './page/Home/HomeMyworkPage';
 import HomePage from './page/Home/HomePage';
-import { SignUpPage } from '@/page/SignUp/SignUpPage';
+import { SignUpPage } from '@/page/SignUp/deprecated/SignUpPage';
 
 import MyPage from './page/MyPage/MyPage';
 import SplashPage from './page/SplashPage';
@@ -15,7 +15,7 @@ import EditApplication from './page/MyPage/EditApplication';
 import { WorkDetailPage } from '@/page/Works/WorkDetailPage';
 import { WorkMainPage } from '@/page/Works/WorkMainPage';
 import { ApplyPage } from '@/page/Apply/ApplyPage';
-import SocialSignUpForm from '@/page/SocialSignup/SocialSignUpForm';
+
 import SocialHomePage from './page/HomeSocial/SocialHomePage';
 import { MatchingApplyPage } from '@/page/Matching/MatchingApplyPage';
 import { MatchingElderPage } from '@/page/Matching/MatchingElderPage';
@@ -28,7 +28,7 @@ import ChatListPage from './page/Chat/ChatListPage';
 import ChatRoomPage from './page/Chat/ChatRoomPage';
 import MatchingStatus from './page/Matching/MatchingStatus';
 import ChatListCaregiver from './page/Chat/ChatListCaregiver';
-import { SignUpForm } from '@/components/SignUp/SignUpForm';
+import { SignUpForm } from '@/components/SignUp/deprecated/SignUpForm';
 
 function App() {
   return (
@@ -41,7 +41,6 @@ function App() {
 
       <Route path="/signup" element={<SignUpPage />} />
       <Route path="/signup/caregiver" element={<SignUpForm />} />
-      <Route path="/signup/social" element={<SocialSignUpForm />} />
 
       <Route path="/home/caregiver" element={<HomePage />} />
       <Route path="/mywork" element={<HomeMyworkPage />} />
