+++ src/components/SignUp/deprecated/AgreeCard.tsx
