+++ src/components/SignUp/deprecated/SignUpForm.tsx
@@ -1,11 +1,11 @@
 import styled from 'styled-components';
-import { Step1 } from '@/page/SignUp/SignStep1';
-import { Step2 } from '@/page/SignUp/SignStep2';
-import { Step3 } from '@/page/SignUp/SignStep3';
-import { Step4 } from '@/page/SignUp/SignStep4';
-import { Step5 } from '@/page/SignUp/SignStep5';
-import { Step6 } from '@/page/SignUp/SignStep6';
-import { Step7 } from '@/page/SignUp/SignStep7';
+import { Step1 } from '@/page/SignUp/deprecated/SignStep1';
+import { Step2 } from '@/page/SignUp/deprecated/SignStep2';
+import { Step3 } from '@/page/SignUp/deprecated/SignStep3';
+import { Step4 } from '@/page/SignUp/deprecated/SignStep4';
+import { Step5 } from '@/page/SignUp/deprecated/SignStep5';
+import { Step6 } from '@/page/SignUp/deprecated/SignStep6';
+import { Step7 } from '@/page/SignUp/deprecated/SignStep7';
 import { useSignUpStep } from '@/hooks/useSignUpStep';
 
 const steps = [Step1, Step2, Step3, Step4, Step5, Step6, Step7];
