--- src/components/SocialSignUp/institutionData.ts
@@ -1,212 +1,0 @@
-export const institutionData = [
-  {
-    institutionId: '11111000006',
-    institutionName: '청운노인요양원',
-    streetAddress: '서울특별시 종로구 비봉길 76',
-    detailAddress: '구기동',
-    openDate: '2008-06-25T00:00:00Z',
-  },
-  {
-    institutionId: '11111000033',
-    institutionName: '청운실버센터',
-    streetAddress: '서울특별시 종로구 자하문로28길 29',
-    detailAddress: '청운동',
-    openDate: '2009-02-27T00:00:00Z',
-  },
-  {
-    institutionId: '11111000056',
-    institutionName: '인지재가센터',
-    streetAddress: '서울특별시 종로구 창신1길 26',
-    detailAddress: '창신동',
-    openDate: '2010-11-25T00:00:00Z',
-  },
-  {
-    institutionId: '11111000062',
-    institutionName: '평창동시니어센터',
-    streetAddress: '서울특별시 종로구 평창문화로 10',
-    detailAddress: '평창동',
-    openDate: '2011-08-16T00:00:00Z',
-  },
-  {
-    institutionId: '11111000068',
-    institutionName: '아름드림한마당요양원',
-    streetAddress: '서울특별시 종로구 평창문화로 70',
-    detailAddress: '평창동',
-    openDate: '2011-08-16T00:00:00Z',
-  },
-  {
-    institutionId: '11111000076',
-    institutionName: '서울여자간호대학교 요양보호센터',
-    streetAddress: '서울특별시 종로구 평창문화로 25',
-    detailAddress: '평창동',
-    openDate: '2012-08-08T00:00:00Z',
-  },
-  {
-    institutionId: '11111000078',
-    institutionName: '세겹장실버홈',
-    streetAddress: '서울특별시 종로구 평창문화로 123',
-    detailAddress: '평창동',
-    openDate: '2013-04-08T00:00:00Z',
-  },
-  {
-    institutionId: '11111000087',
-    institutionName: '해피홈시니어스타워(종로점)',
-    streetAddress: '서울특별시 종로구 평창문화로 123',
-    detailAddress: '평창동',
-    openDate: '2015-12-08T00:00:00Z',
-  },
-  {
-    institutionId: '11111000093',
-    institutionName: '아트온실버홈',
-    streetAddress: '서울특별시 종로구 창신동 25길 8',
-    detailAddress: '창신동',
-    openDate: '2016-03-07T00:00:00Z',
-  },
-  {
-    institutionId: '11111000030',
-    institutionName: '남산실버복지센터',
-    streetAddress: '서울특별시 중구 동호로5길 189',
-    detailAddress: '신당동',
-    openDate: '2010-02-02T00:00:00Z',
-  },
-  {
-    institutionId: '11111000040',
-    institutionName: '구립중구노인요양센터',
-    streetAddress: '서울특별시 중구 필동로 96',
-    detailAddress: '장충동2가',
-    openDate: '2012-09-28T00:00:00Z',
-  },
-  {
-    institutionId: '11114000043',
-    institutionName: '효심원',
-    streetAddress: '서울특별시 중구 청구로17길 98',
-    detailAddress: '신당동',
-    openDate: '2013-11-07T00:00:00Z',
-  },
-  {
-    institutionId: '11121000054',
-    institutionName: '구립용산노인전문요양원',
-    streetAddress: '서울특별시 용산구 효창원로93길 51',
-    detailAddress: '효창동',
-    openDate: '2008-07-09T00:00:00Z',
-  },
-  {
-    institutionId: '11121000092',
-    institutionName: '구립한남노인요양원',
-    streetAddress: '서울특별시 용산구 독서당로 103',
-    detailAddress: '한남동',
-    openDate: '2014-09-01T00:00:00Z',
-  },
-  {
-    institutionId: '11121000102',
-    institutionName: '시립동부노인전문요양센터',
-    streetAddress: '서울특별시 성동구 왕십리로 31',
-    detailAddress: '홍익동',
-    openDate: '2008-06-19T00:00:00Z',
-  },
-  {
-    institutionId: '11122000056',
-    institutionName: '화마음시니어홈',
-    streetAddress: '서울특별시 성동구 마장로125 상왕십리동 진리성회관 5층 6층',
-    detailAddress: '상왕십리동',
-    openDate: '2010-08-12T00:00:00Z',
-  },
-  {
-    institutionId: '11122000085',
-    institutionName: '금호노인요양원',
-    streetAddress: '서울특별시 성동구 금호로 45 B동 201 301호',
-    detailAddress: '금호동4가',
-    openDate: '2010-03-01T00:00:00Z',
-  },
-  {
-    institutionId: '11122000098',
-    institutionName: '이화요양원',
-    streetAddress: '서울특별시 성동구 독서당로 103',
-    detailAddress: '옥수동',
-    openDate: '2010-03-01T00:00:00Z',
-  },
-  {
-    institutionId: '11122000106',
-    institutionName: '서울숲 요양원',
-    streetAddress: '서울특별시 성동구 왕십리로 31',
-    detailAddress: '성수2가1동',
-    openDate: '2015-10-01T00:00:00Z',
-  },
-  {
-    institutionId: '11122000110',
-    institutionName: '서울숲실버요양원',
-    streetAddress: '서울특별시 성동구 독서당로 103',
-    detailAddress: '옥수동',
-    openDate: '2015-10-01T00:00:00Z',
-  },
-  {
-    institutionId: '11121500003',
-    institutionName: '광진구재가요양센터',
-    streetAddress: '서울특별시 광진구 자양로51길 11',
-    detailAddress: '구의동',
-    openDate: '2008-06-25T00:00:00Z',
-  },
-  {
-    institutionId: '11121500005',
-    institutionName: '광진노인요양시설',
-    streetAddress: '서울특별시 광진구 구의사거리 125',
-    detailAddress: '구의동',
-    openDate: '2008-03-03T00:00:00Z',
-  },
-  {
-    institutionId: '11121500009',
-    institutionName: '사랑이있는집노인요양공동생활가정',
-    streetAddress: '서울특별시 광진구 광나루로58길 13-25',
-    detailAddress: '자양동',
-    openDate: '2010-03-01T00:00:00Z',
-  },
-  {
-    institutionId: '11121500014',
-    institutionName: '광진노인전문요양시설',
-    streetAddress: '서울특별시 광진구 자양로58길 12 4층',
-    detailAddress: '자양동',
-    openDate: '2012-06-01T00:00:00Z',
-  },
-  {
-    institutionId: '11121500018',
-    institutionName: '모두솔실버타운',
-    streetAddress: '서울특별시 광진구 자양로 79',
-    detailAddress: '자양동',
-    openDate: '2010-10-01T00:00:00Z',
-  },
-  {
-    institutionId: '11121500036',
-    institutionName: '위카실버타운',
-    streetAddress: '서울특별시 광진구 구의강변로 360',
-    detailAddress: '구의동',
-    openDate: '2011-06-01T00:00:00Z',
-  },
-  {
-    institutionId: '11121500042',
-    institutionName: '성동실버케어요양원',
-    streetAddress: '서울특별시 광진구 아차산로 125',
-    detailAddress: '자양동',
-    openDate: '2015-10-17T00:00:00Z',
-  },
-  {
-    institutionId: '11121500046',
-    institutionName: '광진구립요양원',
-    streetAddress: '서울특별시 광진구 아차산로 613',
-    detailAddress: '자양동',
-    openDate: '2019-10-19T00:00:00Z',
-  },
-  {
-    institutionId: '11121500048',
-    institutionName: '누리봄실버홈',
-    streetAddress: '서울특별시 광진구 구의강변로 305',
-    detailAddress: '구의동',
-    openDate: '2019-10-17T00:00:00Z',
-  },
-  {
-    institutionId: '11121500068',
-    institutionName: '늘푸른실버케어센터1',
-    streetAddress: '서울특별시 광진구 광장동 634',
-    detailAddress: '광장동',
-    openDate: '2019-11-01T00:00:00Z',
-  },
-];
